+++ LeaveManagementApplication.Application/Features/LeaveRequestFeatures/Command/DeleteLeaveRequestCommand.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
+using LeaveManagementApplication.Application.Exceptions;
 using LeaveManagementApplication.Application.Persistance.Contracts;
+using LeaveManagementApplication.Domain.Models;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -31,6 +33,12 @@ namespace LeaveManagementApplication.Application.Features.LeaveRequestFeatures.C
         public async Task<Unit> Handle(DeleteLeaveRequestCommand command, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(command.Id);
+
+
+            if (leaveRequest == null)
+            {
+                throw new NotFoundException(nameof(LeaveRequest), command.Id);
+            }
             await _leaveRequestRepository.delete(leaveRequest);
             return Unit.Value;
         }
