+++ LeaveManagementApplication.Application/Features/LeaveTypeFeatures/Command/DeleteLeaveTypeCommand.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
+using LeaveManagementApplication.Application.Exceptions;
 using LeaveManagementApplication.Application.Persistance.Contracts;
+using LeaveManagementApplication.Domain.Models;
 using MediatR;
 
 namespace LeaveManagementApplication.Application.Features.LeaveTypeFeatures.Command
@@ -30,6 +32,12 @@ namespace LeaveManagementApplication.Application.Features.LeaveTypeFeatures.Comm
         public async Task<Unit> Handle(DeleteLeaveTypeCommand command, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(command.Id);
+
+            if (leaveType == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), command.Id);
+            }
+
             await _leaveTypeRepository.delete(leaveType);
             return Unit.Value;
         }
